refactor: 코드 점검 및 리팩토링

+++ src/main/java/com/nhnacademy/user/dto/response/PointResponse.java
@@ -12,6 +12,8 @@
 
 package com.nhnacademy.user.dto.response;
 
-public record PointResponse(long currentPoint) {
+import java.math.BigDecimal;
+
+public record PointResponse(BigDecimal currentPoint) {
     // 잔액 조회용 응답 DTO
 }
